src/web/app_hybrid.py
#!/usr/bin/env python3
"""
QuantumViz Agent - Hybrid Version
Original app features without problematic dependencies.
"""

from flask import Flask, render_template, request, jsonify, session
from flask_cors import CORS
from flask_limiter import Limiter
from flask_limiter.util import get_remote_address
import json
import os
import sys
import secrets
import re
import time
import numpy as np
import plotly.graph_objects as go
import plotly.utils
from datetime import datetime
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Try to import Google Generative AI
try:
    import google.generativeai as genai
    GEMINI_IMPORTED = True
    
    # Configure Gemini if API key is available
    gemini_api_key = os.getenv('GEMINI_API_KEY')
    if gemini_api_key and gemini_api_key.strip():
        genai.configure(api_key=gemini_api_key)
        GEMINI_AVAILABLE = True
    else:
        GEMINI_AVAILABLE = False
except ImportError:
    GEMINI_IMPORTED = False
    GEMINI_AVAILABLE = False

# Simple quantum circuit simulation class
class QuantumCircuitSimulator:
    """Simplified quantum circuit simulator."""
    
    def __init__(self):
        self.gates = []
        self.n_qubits = 0
    
    def add_gate(self, gate_type, qubit, target=None):
        """Add a gate to the circuit."""
        gate = {'type': gate_type, 'qubit': qubit}
        if target is not None:
            gate['target'] = target
        self.gates.append(gate)
        
        # Update qubit count
        max_qubit = max(qubit, target if target is not None else qubit)
        self.n_qubits = max(self.n_qubits, max_qubit + 1)
    
    def reset(self):
        """Reset the circuit."""
        self.gates = []
        self.n_qubits = 0
    
    def simulate(self, shots=1024):
        """Simulate the quantum circuit."""
        if not self.gates:
            return {}
        
        # Generate realistic quantum results based on gates
        results = {}
        
        # Analyze circuit pattern
        has_hadamard = any(gate['type'] == 'H' for gate in self.gates)
        has_cnot = any(gate['type'] == 'CNOT' for gate in self.gates)
        
        if has_hadamard and has_cnot:
            # Bell state pattern
            results['00'] = np.random.randint(480, 520)
            results['11'] = shots - results['00']
        elif has_hadamard:
            # Superposition
            for i in range(2**self.n_qubits):
                state = format(i, f'0{self.n_qubits}b')
                results[state] = shots // (2**self.n_qubits) + np.random.randint(-20, 21)
        else:
            # Classical state
            results['0' * self.n_qubits] = shots
        
        return results
    
    def get_circuit_string(self):
        """Get a string representation of the circuit."""
        if not self.gates:
            return "Empty circuit"
        
        circuit_str = f"Quantum Circuit ({len(self.gates)} gates, {self.n_qubits} qubits):\n\n"
        
        # Create ASCII quantum circuit
        for q in range(self.n_qubits):
            line = f"q{q} |0⟩ ─"
            for gate in self.gates:
                if gate['qubit'] == q:
                    if gate['type'] == 'H':
                        line += "─[H]─"
                    elif gate['type'] == 'X':
                        line += "─[X]─"
                    elif gate['type'] == 'Y':
                        line += "─[Y]─"
                    elif gate['type'] == 'Z':
                        line += "─[Z]─"
                    elif gate['type'] == 'CNOT' and 'target' in gate:
                        line += "─●───"
                    else:
                        line += "─────"
                elif 'target' in gate and gate['target'] == q:
                    line += "─⊕───"
                else:
                    line += "─────"
            circuit_str += line + "\n"
        
        return circuit_str

def validate_circuit_input(data):
    """Validate circuit input data."""
    if not isinstance(data, dict):
        raise ValueError("Invalid input format")
    
    gate_type = data.get('gate_type')
    if not gate_type:
        raise ValueError("Missing gate_type")
    
    valid_gates = ['H', 'X', 'Y', 'Z', 'CNOT', 'CZ', 'SWAP', 'T']
    if gate_type not in valid_gates:
        raise ValueError(f"Invalid gate type: {gate_type}")
    
    qubit = data.get('qubit')
    if qubit is None or not isinstance(qubit, int) or qubit < 0:
        raise ValueError("Invalid qubit index")
    
    return data

# Initialize Flask app
app = Flask(__name__)

# Enable CORS
CORS(app, origins=["*"], methods=["GET", "POST", "OPTIONS"], 
     allow_headers=["Content-Type", "Authorization"])

# Set secret key
app.secret_key = os.getenv('FLASK_SECRET_KEY', secrets.token_hex(32))

# Configure rate limiting
limiter = Limiter(
    app=app,
    key_func=get_remote_address,
    default_limits=["200 per day", "50 per hour"],
    storage_uri="memory://"
)

# Initialize quantum simulator
quantum_sim = QuantumCircuitSimulator()

@app.route('/')
def index():
    """Main page."""
    return jsonify({
        'message': 'QuantumViz Agent - Hybrid Version',
        'status': 'running',
        'features': ['quantum_simulation', 'ai_explanations', 'visualization']
    })

@app.route('/api/health')
def health_check():
    """Health check endpoint."""
    return jsonify({
        'status': 'healthy',
        'service': 'QuantumViz Agent API',
        'version': '1.0.0',
        'aws_available': False,
        'gemini_available': GEMINI_AVAILABLE
    })

@app.route('/api/add_gate', methods=['POST', 'OPTIONS'])
@limiter.limit("30 per minute")
def add_gate():
    """Add quantum gate to circuit."""
    if request.method == 'OPTIONS':
        return '', 200
    
    try:
        data = request.json
        validated_data = validate_circuit_input(data)
        
        quantum_sim.add_gate(
            validated_data['gate_type'],
            validated_data['qubit'],
            validated_data.get('target')
        )
        
        return jsonify({
            'status': 'success',
            'circuit': quantum_sim.get_circuit_string(),
            'gate_count': len(quantum_sim.gates)
        })
    
    except Exception as e:
        return jsonify({
            'status': 'error',
            'message': str(e)
        }), 400

@app.route('/api/simulate', methods=['POST', 'OPTIONS'])
@limiter.limit("10 per minute")
def simulate():
    """Run quantum simulation."""
    if request.method == 'OPTIONS':
        return '', 200
    
    try:
        if not quantum_sim.gates:
            return jsonify({
                'status': 'error',
                'message': 'No gates in circuit'
            }), 400
        
        # Run simulation
        results = quantum_sim.simulate()
        
        # Create visualization
        states = list(results.keys())
        counts = list(results.values())
        
        fig = go.Figure(data=[
            go.Bar(x=states, y=counts, 
                   marker_color=['#FF6B6B', '#4ECDC4', '#45B7D1', '#96CEB4'])
        ])
        
        fig.update_layout(
            title='Quantum Measurement Results',
            xaxis_title='Quantum States',
            yaxis_title='Probability',
            height=400,
            paper_bgcolor='#212121',
            plot_bgcolor='#212121',
            font={'color': '#FFFFFF', 'family': 'JetBrains Mono'}
        )
        
        visualization = json.dumps(fig, cls=plotly.utils.PlotlyJSONEncoder)
        
        return jsonify({
            'status': 'success',
            'results': results,
            'visualization': visualization,
            'gate_count': len(quantum_sim.gates),
            'qubit_count': quantum_sim.n_qubits
        })
    
    except Exception as e:
        return jsonify({
            'status': 'error',
            'message': str(e)
        }), 400

@app.route('/api/reset', methods=['POST', 'OPTIONS'])
def reset():
    """Reset quantum circuit."""
    if request.method == 'OPTIONS':
        return '', 200
    
    quantum_sim.reset()
    return jsonify({
        'status': 'success',
        'message': 'Circuit reset'
    })

@app.route('/api/ai/explain', methods=['POST', 'OPTIONS'])
@limiter.limit("5 per minute")
def ai_explain():
    """Get AI explanation of circuit."""
    if request.method == 'OPTIONS':
        return '', 200
    
    try:
        if not quantum_sim.gates:
            return jsonify({
                'status': 'error',
                'message': 'No circuit to explain'
            }), 400
        
        # Try Gemini AI first
        if GEMINI_AVAILABLE:
            try:
                model = genai.GenerativeModel('gemini-pro')
                
                circuit_description = f"Quantum circuit with {len(quantum_sim.gates)} gates: "
                circuit_description += ", ".join([f"{gate['type']} on qubit {gate['qubit']}" + 
                                                (f" targeting qubit {gate['target']}" if 'target' in gate else "")
                                                for gate in quantum_sim.gates])
                
                prompt = f"""Explain this quantum circuit in simple terms for educational purposes:

{circuit_description}

Provide:
1. What this circuit does
2. The quantum concepts involved  
3. Expected measurement results
4. Real-world applications

Keep it beginner-friendly but technically accurate."""
                
                response = model.generate_content(prompt)
                explanation = response.text
                
                return jsonify({
                    'status': 'success',
                    'explanation': explanation,
                    'model_used': 'gemini-pro'
                })
            
            except Exception as e:
                print(f"Gemini AI error: {e}")
                # Fall through to fallback explanation
        
        # Fallback explanation
        gate_types = [gate['type'] for gate in quantum_sim.gates]
        
        explanation = "🤖 QuantumViz AI Circuit Analysis\n\n"
        explanation += f"📊 Circuit Statistics:\n"
        explanation += f"   • {len(quantum_sim.gates)} quantum gates\n"
        explanation += f"   • {quantum_sim.n_qubits} qubits involved\n\n"
        
        explanation += "🔬 Quantum Operations:\n"
        
        if 'H' in gate_types and 'CNOT' in gate_types:
            explanation += "🎯 ENTANGLEMENT DETECTED!\n"
            explanation += "   • Creates Bell state: |00⟩ + |11⟩\n"
            explanation += "   • Demonstrates quantum correlation\n\n"
        
        if 'H' in gate_types:
            explanation += "✓ Hadamard Gates: Create superposition\n"
        if 'CNOT' in gate_types:
            explanation += "✓ CNOT Gates: Create entanglement\n"
        if 'X' in gate_types:
            explanation += "✓ Pauli-X Gates: Quantum NOT operation\n"
        
        explanation += "\n📈 Expected Results:\n"
        if 'H' in gate_types and 'CNOT' in gate_types:
            explanation += "   • 50% |00⟩, 50% |11⟩ (entangled)\n"
        elif 'H' in gate_types:
            explanation += "   • Equal superposition of all states\n"
        
        explanation += "\n💡 Applications: Quantum cryptography, teleportation, computing"
        
        return jsonify({
            'status': 'success',
            'explanation': explanation,
            'model_used': 'fallback'
        })
    
    except Exception as e:
        return jsonify({
            'status': 'error',
            'message': str(e)
        }), 400

if __name__ == '__main__':
    port = int(os.getenv('PORT', 5000))
    host = '0.0.0.0'
    debug_mode = os.getenv('FLASK_ENV', 'development') == 'development'
    app.run(host=host, port=port, debug=debug_mode)