requirements_original.txt
# QuantumViz Agent - Working Dependencies for Python 3.11

# Web Framework
flask==2.3.3
flask-cors==4.0.0
flask-limiter==3.5.0

# Essential packages only
numpy>=1.21.0
plotly>=5.0.0
requests>=2.25.0
python-dotenv>=0.19.0

# AI Services (optional)
google-generativeai>=0.3.0

# AWS Services (optional)
boto3>=1.20.0